Remove auto-loading of cards based on ID
Removed cardent (ensure_entity) usage for actions which provide card as
ID.
ensure_entity would convert this to an object using cardent.load
transparently. This caused a problem when zones are injected into
entity.make$ _after_ ensure_entity is called.

It also turned out to be totally unused here as all actions used the ID
as a starting point anyway, so it's a bit of an optimisation.

Note: This is a breaking change as callers can no longer pass a card
object and expect it to work (for the 'children' action). The card ID
must now be passed as the 'card' arg.