Change id$ usage to just "id" in cmd_relate
Our seneca wrappers are intercepting entity saves and creating a new
message with a zone field for seneca to create the entity in a specific
zone. Part of this involves taking the data from the original message
and stripping seneca fields out with .data$(). id$ was being stripped
out which resulted in seneca cards being created with a different id to
the specialized card.
This change retains the data and, seemingly, the data.